requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
future==1.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
netmiko==4.5.0
ntc_templates==7.7.0
paramiko==3.5.1
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scp==0.15.0
simple-websocket==1.1.0
six==1.17.0
textfsm==1.1.3
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
